refactor(InputFile): client専用のDownloadAnchorButtonをclient/へ切り出しparts.tsxから'use client'を除去する
DownloadAnchorButtonはuseState/useEffectでBlob URLを生成するclient専用
処理のため、client/へ切り出し'use client'を付与する。

parts.tsx自体はindex.tsから公開されておらず、唯一の利用元である
InputFileNative.tsx/InputFileMultiplyAppendable.tsxは共に'use client'
を持つため、parts.tsxからの'use client'除去は安全（常にClient
Componentのサブツリー内でのみレンダーされる）。
