+++ packages/smarthr-ui/src/components/InputFile/client/index.ts
@@ -1,0 +1,1 @@
+export { DownloadAnchorButton } from './DownloadAnchorButton'
