+++ packages/smarthr-ui/src/components/InputFile/parts.tsx
@@ -1,11 +1,12 @@
-'use client'
-
-import { type FC, type MouseEvent, type ReactNode, memo, useEffect, useState } from 'react'
+import { type FC, type MouseEvent, type ReactNode, memo } from 'react'
 
 import { Localizer } from '../../intl'
-import { AnchorButton, Button } from '../Button'
-import { FaFileArrowDownIcon, FaFileLinesIcon, FaFolderOpenIcon, FaTrashCanIcon } from '../Icon'
+import { Button } from '../Button'
+import { FaFileLinesIcon, FaFolderOpenIcon, FaTrashCanIcon } from '../Icon'
 import { VisuallyHiddenText } from '../VisuallyHiddenText'
+
+import { DownloadAnchorButton } from './client'
+import { FILE_NAME_BUTTON_CLASSNAME, PREVIEW_BUTTON_CLASSNAME } from './style'
 
 export const StyledFaFolderOpenIcon = memo<{ className: string }>(({ className }) => (
   <span className={className}>
@@ -18,10 +19,6 @@ export const LabelRender = memo<{ id: string; label: ReactNode }>(({ id, label }
     {label}
   </span>
 ))
-
-const FILE_NAME_BUTTON_CLASSNAME =
-  'smarthr-ui-InputFile-fileName shr-justify-start shr-min-w-0 shr-break-all shr-whitespace-normal shr-text-left'
-const PREVIEW_BUTTON_CLASSNAME = `${FILE_NAME_BUTTON_CLASSNAME} shr-p-0 shr-font-normal shr-text-link`
 
 const PreviewButton: FC<{
   file: File
@@ -43,38 +40,6 @@ const PreviewButton: FC<{
     </VisuallyHiddenText>
   </Button>
 )
-
-const DownloadAnchorButton: FC<{ file: File }> = ({ file }) => {
-  const [href, setHref] = useState('')
-
-  useEffect(() => {
-    const url = URL.createObjectURL(file)
-    setHref(url)
-
-    return () => {
-      URL.revokeObjectURL(url)
-    }
-  }, [file])
-
-  return (
-    <AnchorButton
-      href={href}
-      download={file.name}
-      variant="text"
-      className={PREVIEW_BUTTON_CLASSNAME}
-      prefix={<FaFileArrowDownIcon className="shr-shrink-0" />}
-    >
-      <span aria-hidden="true">{file.name}</span>
-      <VisuallyHiddenText>
-        <Localizer
-          id="smarthr-ui/InputFile/downloadLabel"
-          defaultText="{fileName}をダウンロード"
-          values={{ fileName: file.name }}
-        />
-      </VisuallyHiddenText>
-    </AnchorButton>
-  )
-}
 
 type FileListItemProps = {
   file: File
